Update PDBtool.py
Added the while loop for the command line to display "Enter command " after introduction. Included placeholder code in each function to display a string when  a command is entered. Please remove and replace with your own code when you are able to. Also notice the arguments passed in each function.